+++ itest/rust/src/packed_array_test.rs
@@ -4,7 +4,7 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-use crate::{expect_panic, itest};
+use crate::framework::{expect_panic, itest};
 use godot::builtin::{PackedByteArray, PackedFloat32Array, PackedStringArray};
 
 #[itest]
