+++ itest/rust/src/plane_test.rs
@@ -4,7 +4,7 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-use crate::itest;
+use crate::framework::itest;
 
 use godot::builtin::inner::InnerPlane;
 use godot::builtin::math::{assert_eq_approx, ApproxEq};
