+++ itest/rust/src/string/godot_string_test.rs
@@ -6,7 +6,7 @@
 
 use std::collections::HashSet;
 
-use crate::itest;
+use crate::framework::itest;
 use godot::builtin::GodotString;
 
 // TODO use tests from godot-rust/gdnative
