+++ itest/rust/src/transform2d_test.rs
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
-use crate::itest;
+use crate::framework::itest;
 
 use godot::prelude::{inner::InnerTransform2D, *};
 use godot::private::class_macros::assert_eq_approx;
